Update StepController test organization
* Split tests of different files into different files.
* Merged factory tests into main tests.
