+++ tests/Unit/Time/CMakeLists.txt
@@ -9,13 +9,13 @@ set(LIBRARY_SOURCES
   Test_PastTime.cpp
   Test_Slab.cpp
   Test_SpecifiedSlabs.cpp
-  Test_StepControllers.cpp
   Test_Time.cpp
   Test_TimeId.cpp
   )
 
 add_subdirectory(Actions)
 add_subdirectory(StepChoosers)
+add_subdirectory(StepControllers)
 add_subdirectory(TimeSteppers)
 
 add_test_library(
