+++ tests/Unit/Time/StepControllers/CMakeLists.txt
@@ -1,0 +1,10 @@
+# Distributed under the MIT License.
+# See LICENSE.txt for details.
+
+set(LIBRARY_SOURCES
+  ${LIBRARY_SOURCES}
+  StepControllers/Test_BinaryFraction.cpp
+  StepControllers/Test_FullSlab.cpp
+  StepControllers/Test_SimpleTimes.cpp
+  StepControllers/Test_SplitRemaining.cpp
+  PARENT_SCOPE)
